add option to combine two brush settings
